Added functions set and set helper.
Set is used to set a cell to the desired state while maintaining its Ground. set itself just grabs the row in which the cell is located in and then passes it to setHelper which handles selecting the correct cell from the row that is given and returning a new row with the old cell replaced with a new cell containing the same ground but has a different state.